Add SSH agent autoload in bashrc for physical TTYs
Because of the current state of OSX SSH and it not supporting the 1800+ line
patch for OSXs keychain. This is my workaround to keeping SSH up to date for
security reasons. Also being able to support the new ECDSA (curve25519) and
Edward Tiwsted Curves (ed25519), which are way more performant and secure than
RSA. At least against 3 letter government organizations.

This is also in response to Apple not keeping ssh on their BSD fork Darwin up
to date when it comes to SSH and openssl or libressl or other crypto libs.